test: split agent provider type tests
